UI: new PIN flow
Previously, when user clicked to scan the card, it was scanned a first time to get status (need setup or ready),
then scan was interrupted to let user enter its PIN (when card ready) or provide a new PIN.
After that, user had to rescan the  card to display its content.

With new PIN flow, the user is first asked to provide PIN code, then the card is scaned only once, which is faster and more convenient.